gpkg2: bi_biotopes

+++ gpkg2.sh
@@ -594,7 +594,7 @@ fi
 #                                                                          #
 # ======================================================================== #
 
-# 25'28"
+# 17'25"
 
 # Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
 vconf=`grep "$t_gpk" "$fconf"`
@@ -860,6 +860,52 @@ then
 
 	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
 	cp_gpk "$typ01" "$cv_gpk"
+	msg " - ${typ01}"
+	rm "$f02" 2> /dev/null
+fi
+
+# =================================== #
+#                                     #
+# 17. bi_biotopes (biotopa / biotopo) #
+#                                     #
+# =================================== #
+
+# 9"
+
+# Konfigurazio-fitxategia irakurri / Leer el fichero de configuración
+vconf=`grep "$bi_gpk" "$fconf"`
+IFS='|' read -a aconf <<< "$vconf"
+typ01="${aconf[0]}"
+gpk01="${aconf[1]}"
+des01="${bi_des[0]} - ${bi_des[1]} - ${bi_des[2]}"
+if [ "$bi_gpk" = "$gpk01" ] && ([ $typ01 = "1" ] || [ "$typ01" = "2" ])
+then
+	let i1=$i1+1
+	msg "${i1}/${nf}: $(date '+%Y-%m-%d %H:%M:%S') - $gpk01 - ${bi_des[0]}\c"
+	f01="${tmpd}/${bi_gpk}_01.gpkg"
+	c01="${tmpd}/${bi_gpk}_01.csv"
+	f02="${tmpd}/${bi_gpk}.gpkg"
+
+	# Oinarrizko datuak / Datos básicos
+	rm "$f01" 2> /dev/null
+	ogr2ogr -f "GPKG" -s_srs "EPSG:25830" -t_srs "EPSG:25830" "$f01" OCI:${con}:${tpl} -nln "$bi_gpk" -lco DESCRIPTION="$des01" -sql "$bi_sql_01"
+
+	# more_info
+	rm "$c01" 2> /dev/null
+	sql_more_info2 "$c01" "$bi_sql_02"
+	ogr2ogr -f "GPKG" -update -s_srs "EPSG:25830" -t_srs "EPSG:25830" -nln "${bi_gpk}_more_info" -lco DESCRIPTION="${des01} more info" "$f01" "$c01"
+	rm "$c01" 2> /dev/null
+
+	# Behin betiko GPKGa / GPKG definitivo
+	rm "$f02" 2> /dev/null
+	ogr2ogr -f "GPKG" -s_srs "EPSG:25830" -t_srs "EPSG:25830" -nln "$bi_gpk" -lco DESCRIPTION="$des01" -sql "$bi_sql_03" "$f02" "$f01"
+	rm "$f01" 2> /dev/null
+
+	# Eremuak berrizendatu / Renombrar campos
+	rfl "$f02" "$bi_gpk"
+
+	# Garapenera edo ekoizpenera kopiatu / Copiar a desarrollo o a producción
+	cp_gpk "$typ01" "$bi_gpk"
 	msg " - ${typ01}"
 	rm "$f02" 2> /dev/null
 fi
